resolving user logic error

+++ src/organizer/organizer.service.ts
@@ -1,8 +1,9 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Organizer } from './entities/organizer.entity';
 import { CreateOrganizerDto } from './dto/create-organizer.dto';
+import { User } from 'src/users/entities/users.entity';
 
 @Injectable()
 export class OrganizerService {
@@ -10,55 +11,80 @@ export class OrganizerService {
   constructor(
     @InjectRepository(Organizer)
     private organizerRepository: Repository<Organizer>,
+    @InjectRepository(User)
+    private userRepository: Repository<User>
   ) { }
 
-  async findAll(): Promise<Organizer[]> {
-    try {
-      const organizers = await this.organizerRepository.find({ relations: ['organizerCategory'] });
-      return organizers;
-    } catch (err) {
-      throw new Error(`Failed to retrieve all organizers: ${err.message}`);
-    }
+  async findAll(userId: number): Promise<Organizer[]> {
+    const user = await this.userRepository.findOne({ where: { id: userId } })
+
+    if (user == null) throw new NotFoundException('user not found')
+
+    if (user.isAdmin != true) throw new BadRequestException('user is not admin')
+
+    const organizers = await this.organizerRepository.find({ relations: ['organizerCategory'] });
+    return organizers;
+
+
   }
 
-  async findById(id: number): Promise<Organizer> {
-    try {
-      const organizer = await this.organizerRepository.findOne({ where: { id: id }, relations: ['organizerCategory'] });
-      if (!organizer) throw new Error(`Organizer with id ${id} not found`);
-      return organizer;
-    } catch (err) {
-      throw new Error(`Failed to retrieve organizer with id ${id}: ${err.message}`);
-    }
+  async findById(id: number, userId: number): Promise<Organizer> {
+    const user = await this.userRepository.findOne({ where: { id: userId } })
+
+    if (user == null) throw new NotFoundException('user not found')
+
+    if (user.isAdmin != true) throw new BadRequestException('user is not admin')
+    const organizer = await this.organizerRepository.findOne({ where: { id: id }, relations: ['organizerCategory'] });
+
+    if (!organizer) throw new NotFoundException(`Organizer with id ${id} not found`);
+    return organizer;
+
   }
 
-  async create(organizer: CreateOrganizerDto): Promise<Organizer> {
-    try {
-      const newOrganizer = await this.organizerRepository.save(organizer);
-      return newOrganizer;
-    } catch (err) {
-      throw new Error(`Failed to create organizer: ${err.message}`);
-    }
+  async create(organizer: CreateOrganizerDto, userId: number): Promise<Organizer> {
+    const user = await this.userRepository.findOne({ where: { id: userId } })
+
+    if (user == null) throw new NotFoundException('user not found')
+
+    if (user.isAdmin != true) throw new BadRequestException('user is not admin')
+
+    const newOrganizer = await this.organizerRepository.save(organizer);
+    return newOrganizer;
+
   }
 
-  async update(id: number, organizer: CreateOrganizerDto): Promise<Organizer> {
-    try {
-      await this.organizerRepository.update(id, organizer);
-      const updatedOrganizer = await this.organizerRepository.findOne({ where: { id: id }, relations: ['organizerCategory'] });
-      if (!updatedOrganizer) throw new Error(`Organizer with id ${id} not found after update`);
-      return updatedOrganizer;
-    } catch (err) {
-      throw new Error(`Failed to update organizer with id ${id}: ${err.message}`);
-    }
+  async update(id: number, organizer: CreateOrganizerDto, userId: number): Promise<Organizer> {
+
+    const user = await this.userRepository.findOne({ where: { id: userId } })
+
+    if (user == null) throw new NotFoundException('user not found')
+
+    if (user.isAdmin != true) throw new BadRequestException('user is not admin')
+
+    await this.organizerRepository.update(id, organizer);
+
+    const updatedOrganizer = await this.organizerRepository.findOne({ where: { id: id }, relations: ['organizerCategory'] });
+
+    if (!updatedOrganizer) throw new Error(`Organizer with id ${id} not found after update`);
+
+    return updatedOrganizer;
+
   }
 
-  async delete(id: number): Promise<Organizer> {
-    try {
-      const organizer = await this.organizerRepository.findOne({ where: { id: id }, relations: ['organizerCategory'] });
-      if (!organizer) throw new Error(`Organizer with id ${id} not found`);
-      const deleted = await this.organizerRepository.remove(organizer);
-      return deleted
-    } catch (err) {
-      throw new Error(`Failed to delete organizer with id ${id}: ${err.message}`);
-    }
+  async delete(id: number, userId: number): Promise<Organizer> {
+    const user = await this.userRepository.findOne({ where: { id: userId } })
+
+    if (user == null) throw new NotFoundException('user not found')
+
+    if (user.isAdmin != true) throw new BadRequestException('user is not admin')
+
+    const organizer = await this.organizerRepository.findOne({ where: { id: id }, relations: ['organizerCategory'] });
+
+    if (!organizer) throw new Error(`Organizer with id ${id} not found`);
+
+    const deleted = await this.organizerRepository.remove(organizer);
+    
+    return deleted
+
   }
 }
